Do not get the file_locator service alias
As the aliased service Symfony\Component\HttpKernel\Config\FileLocator
is private, sepending on your Symfony version, it's deprecated or not
possible at all to get it from the container.

As Dependency Injection is no solution here, the next best thing to do
seems to instantiate the FileLocator by ourselves. This might be an
issue in exotic cases when one manipulated the service or if it's
constructor changes, but that is a problem for another day.